extract RSValue as small CMake module
