+++ CMakeLists.txt
@@ -237,6 +237,7 @@ add_subdirectory(src/geometry)
 add_subdirectory(src/inverted_index)
 add_subdirectory(src/module-init)
 add_subdirectory(src/wildcard)
+add_subdirectory(src/value)
 add_subdirectory(src/util/arr)
 add_subdirectory(src/util/hash)
 add_subdirectory(src/util/mempool)
@@ -285,6 +286,7 @@ set(FINAL_OBJECTS
     $<TARGET_OBJECTS:inverted_index>
     $<TARGET_OBJECTS:mempool>
     $<TARGET_OBJECTS:module_init>
+    $<TARGET_OBJECTS:value>
     $<TARGET_OBJECTS:wildcard>
     $<TARGET_OBJECTS:rscore>
     $<TARGET_OBJECTS:rmutil>
@@ -318,6 +320,7 @@ target_link_libraries(redisearch
         inverted_index
         mempool
         module_init
+        value
         wildcard
         uv_a
         ${BINDIR}/redisearch_rs/libredisearch_rs.a
